refactor(scoring): split scoring into src/utils/scoring modules
`getPlayerScores.ts` held 593 lines covering seven unrelated jobs, so any change
to one of them meant reading all of it. Each job is now its own module and
`getPlayerScores` is a 40-line sequence of them.

* `readFileToBuffer` moves to `src/utils/readFileToBuffer.ts`. It wraps
  `FileReader` and has nothing to do with scoring.
* The sort comparator becomes a list of tiers, one function each. Every tier
  keeps its original comparison, name ordering included, so the rank order does
  not move.
* `getPickResults` and `getPlayerScores` tests move next to their modules. No
  case bodies change.
