+++ src/components/RakSadness.tsx
@@ -19,7 +19,8 @@ import { RakMadnessScores } from "../types/RakMadnessScores";
 import buildSpreadsheetBuffer from "../utils/buildSpreadsheetBuffer";
 import getClasses from "../utils/getClasses";
 import getLeagueInfo from "../utils/getLeagueInfo";
-import { getPlayerScores, readFileToBuffer } from "../utils/getPlayerScores";
+import { readFileToBuffer } from "../utils/readFileToBuffer";
+import { getPlayerScores } from "../utils/scoring/getPlayerScores";
 import Footer from "./footer/Footer";
 import LogoButton from "./navbar/LogoButton/LogoButton";
 import Navbar from "./navbar/Navbar";
