+++ src/utils/CLAUDE.md
@@ -1,4 +1,8 @@
 # src/utils
 
-getLeagueInfo/getLeagueResults: ESPN API fetch, calendar/week mapping. getPlayerScores: spread-adjusted pick scoring, knockout detection, tiebreaker logic. buildSpreadsheetBuffer: xlsx-js-style workbook export. getClasses: conditional className join. rangeWithPrefix: labeled index arrays (C1, C2…).
+getLeagueInfo/getLeagueResults: ESPN API fetch, calendar/week mapping. buildSpreadsheetBuffer: xlsx-js-style workbook export. getClasses: conditional className join. rangeWithPrefix: labeled index arrays (C1, C2…). readFileToBuffer: FileReader wrapper for an uploaded spreadsheet.
 leagueResultFixtures: `finalGame`/`upcomingGame` builders shared by the tests here.
+
+## Subdirectories
+
+- [`scoring/`](scoring/CLAUDE.md) — picks to scoreboard: parsing, scoring, knockouts
